backend/banking_api/requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
typing_extensions
validate-docbr
gunicorn
